ar71xx: fix 100/10mbps ethernet link issues on mynet range extender
The mynet range extender hardware is suffering from ethernet
link loss when booting with a recent openwrt image. This only
happens on 100mbps links, with 1Gbps speed the link was fine.

The cause of the problem is that the AR8035 PHY (aka F1E)
requires turning on and off the special TX delay setting
depending on the speed of the link.

The 10mbps mode only needed the proper pll value, which was
extracted from the vendor code.

Reported-by: Pascal Paradis
Signed-off-by: Christian Lamparter <chunkeey@googlemail.com>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@45954 3c298f89-4303-0410-b956-a3cf2f4a3e73