CC: mac80211: disable ipw2x00 for big endian targets
The ipw2x00 drivers assume that the system they are running is little
endian, and access everything in native byte order. When run on a big
endian system, everything breaks apart.

Since fixing this is non trivial on a first glance, disable them for
big endian targets.

Backport of r46708.

Signed-off-by: Jonas Gorski <jogo@openwrt.org>

git-svn-id: svn://svn.openwrt.org/openwrt/branches/chaos_calmer@46710 3c298f89-4303-0410-b956-a3cf2f4a3e73