avanta: Add Marvell RTC to all machines with nothing else
The Avanta SoCs all have an onchip RTC that can hold the time
over e.g. a reboot which will help if no NTP servers are available.

Create a kernel module package for the Marvell RTC, and add it to
all Avanta devices that do not have their own discrete
battery-backed RTC. Adding it to platforms with a proper RTC
is just surplus.

All Avantas have at least one RTC so add RTC to the features
list for Avanta as well.